scripts/generate_data: Rework oracle to avoid generating forbidden demonstrations
A flaw with the oracle was that it could generate demonstrations of things
in splits B, C, D, E and F. Even though we weren't really trying to optimize
performance on those splits, it makes benchmarks on them problematic, since
we technically leak the solution in the demonstrations.

This change reworks the oracle so that generating those demonstrations is
not possible in the training set, unless it is the only viable demonstration.
Instead we generate demonstrations of going to *different* objects and try
to learn from those.

In the test data, we allow generating demonsrtations of other splits other
than the one being tested.