generate_data: Add --world-encoding-scheme option with choices sequence/all
In some datasets like ReaSCAN, you can have two objects occupying the same
grid cell. This doesn't work in the encoding scheme that gives the entire
grid as the observation as opposed to just the non-empty cells.

To get around this, we have a new encoding scheme called "sequence"
which just gives the list of objects in the world and their position encodings.

This also means that we don't add the position encodings after the fact
anymore, instead they're added during the encoding of parse_sparse_situation