generate_data_imagine_trajectories: Use sampling-based balancing
As opposed to limiting the dataset. The reason why limiting is a
bad idea is that we might accidentally introduce a systematic shift
(eg, we limit the number of times we see "push a big circle" ->
we never see the instruction with green circles, so we don't know
how to generate "push a big circle" when green circles are the biggest
circle.