meta_encdec_big_img_transformer: Fix computation of reselection_idx
Previously we computed the target locations of the padding and
non-padding tokens, then passed that to torch.gather. That's actually
wrong - because torch.gather will fetch from the target locations as
opposed to moving the padding to the target locations.

To fetch from the correct places, we argsort the target location tensor.
This means that we fetch all the non-padding to the beginning and
then fetch all the padding at the end.